Performance Notes (no-ci)
After reviewing the per-frame JS heap size between master and dots, we
see a 91.7091% decrease in the amount of memory allocated per frame with
this PR.

+--------+------------------------+
| Branch | JS Heap Growth / Frame |
+--------+------------------------+
| master |         847384         |
|  dots  |          70256         |
+--------+------------------------+
| % diff |            -91.7091%   |
+--------+------------------------+

Furthermore, these changes move the primary render frame updates to a
worker! This frees the main-thread for other user tasks like rendering
the blog calender.