DOTS: update instead of rebuild :zap:
After watching the QuadTree with the debugger enabled, it finally dawned
on me that a majority of the Quads will be staying the same from frame
to frame, After running a few tests, I found that between 0 and 10 Quads
would be updated on any given frame (during normal operation).  Thus, to
save a bit of time re-building the tree constantly, this change adds the
ability to update points within the tree.

After reviewing the logic, there are two parts to this operation:
  1) find the associated quad for a given dot
  2) remove the dot from it's current quad
  3) place the dot back into the tree

To be able to find a given quad from a dot, I added a pointer to the dot
object that points to the quad that it's currently housed in.  This is
used when dots are being iterated from the dots array and not as
children of the quad tree.  A future refactor of this could move to a
tree traversal based iteration of all the nodes instead of maintaining a
separate array of all the dots on the screen.

To remove the dot from it's current node, `update` first removes the dot
from it's points array (a constant sized buffer) and then determines if
the node is empty.  If the quad is empty, it notifies it's parent of the
situation, which can then merge it's 4 child quads into itself if the
quad capacity is not exceeded when doing so.  This also allows all the
children to be recycled for future use.  A few edge cases here that were
just ignored and the process exits out early.  A side effect of this is
that the quads now hold full dots rather than just their position so as
quads are rebalanced, we can update dot's quad pointers appropriately.

To place the dot back into the tree, I just did the full quad insert
starting from the root of the tree.  An alternative approach could have
the quad call out to it's parent until the quad contains the dot, and
then adding the dot to that quad (recursively).  Given the shallow depth
of the trees in this scenario (from 4-7) depending on resolution, the
overall distance traveled by these updates would be the same in big-O
notation.  A few nuances here about the constant factor being removed,
but since data locality isn't a problem (often seen in DBs), a re-insert
from the root is fine.